l03/requirements.txt
jupyter==1.0.0
matplotlib==3.2.0
numpy==1.18.1
pyro-ppl==1.3.0
scikit-learn==0.22.2
torch==1.2.0
tqdm==4.43.0
